chore(server): move organismes.actions to organismes folder
