+++ server/src/common/actions/organismes/organismes.actions.js
@@ -1,16 +1,16 @@
 import { ObjectId } from "mongodb";
-import { getMetiersBySirets } from "../apis/apiLba.js";
-import { organismesDb } from "../model/collections.js";
-import { defaultValuesOrganisme, validateOrganisme } from "../model/next.toKeep.models/organismes.model.js";
-import { buildAdresseFromApiEntreprise } from "../utils/adresseUtils.js";
-import { buildTokenizedString } from "../utils/buildTokenizedString.js";
-import { generateKey } from "../utils/cryptoUtils.js";
-import { buildAdresseFromUai } from "../utils/uaiUtils.js";
-import { siretSchema } from "../utils/validationUtils.js";
-import { mapFiabilizedOrganismeUaiSiretCouple } from "./engine/engine.organismes.utils.js";
-import { createPermission, hasPermission } from "./permissions.actions.js";
-import { findRolePermissionById } from "./roles.actions.js";
-import { getUser } from "./users.actions.js";
+import { getMetiersBySirets } from "../../apis/apiLba.js";
+import { organismesDb } from "../../model/collections.js";
+import { defaultValuesOrganisme, validateOrganisme } from "../../model/next.toKeep.models/organismes.model.js";
+import { buildAdresseFromApiEntreprise } from "../../utils/adresseUtils.js";
+import { buildTokenizedString } from "../../utils/buildTokenizedString.js";
+import { generateKey } from "../../utils/cryptoUtils.js";
+import { buildAdresseFromUai } from "../../utils/uaiUtils.js";
+import { siretSchema } from "../../utils/validationUtils.js";
+import { mapFiabilizedOrganismeUaiSiretCouple } from "../engine/engine.organismes.utils.js";
+import { createPermission, hasPermission } from "../permissions.actions.js";
+import { findRolePermissionById } from "../roles.actions.js";
+import { getUser } from "../users.actions.js";
 
 /**
  * Méthode de création d'un organisme qui applique en entrée des filtres / rejection
