+++ server/src/http/middlewares/requireJwtAuthentication.js
@@ -3,7 +3,7 @@ import { Strategy as JwtStrategy, ExtractJwt } from "passport-jwt";
 import config from "../../config.js";
 import { tdbRoles } from "../../common/roles.js";
 import { getUserLegacy } from "../../common/actions/legacy/users.legacy.actions.js";
-import { findOrganismeByUai } from "../../common/actions/organismes.actions.js";
+import { findOrganismeByUai } from "../../common/actions/organismes/organismes.actions.js";
 
 export default () => {
   const findUserOrCfa = async (usernameOrUai) => {
