+++ server/tests/integration/common/actions/organismes.actions.test.js
@@ -6,7 +6,7 @@ import {
   createOrganisme,
   findOrganismeById,
   updateOrganisme,
-} from "../../../../src/common/actions/organismes.actions.js";
+} from "../../../../src/common/actions/organismes/organismes.actions.js";
 import { buildTokenizedString } from "../../../../src/common/utils/buildTokenizedString.js";
 import { fiabilisationUaiSiretDb } from "../../../../src/common/model/collections.js";
 import { FIABILISATION_MAPPINGS } from "../../../../src/jobs/fiabilisation/uai-siret/create-fiabilisation-uai-siret-mapping/mapping.js";
